Move `localVariable` up to `MethodElementBuilder`
